requirements.txt
Pillow==5.3.0
imutils==0.5.1
numpy==1.15.4
pydub==0.23.0
